More conversions from int -> long

+++ Herd.Business/Models/Commands/GetRecentPostsCommand.cs
@@ -4,8 +4,8 @@
     {
         public bool IncludeInReplyToPost { get; set; }
         public bool IncludeReplyPosts { get; set; }
-        public int? SincePostID { get; set; }
-        public int? MaxPostID { get; set; }
+        public long? SincePostID { get; set; }
+        public long? MaxPostID { get; set; }
         public int MaxCount { get; set; } = 30;
     }
 }
